Intuitionize section "Infinite primes theorem" (#4326)
* Move nnmindc and nnminle earlier in iset.mm

Make a section "Decidable sets of integers"

* Add nnwodc and uzwodc to iset.mm

This is nnwo and uzwo from set.mm but where the set of
integers is decidable.  The proofs use existing iset.mm theorems
concerning decidable sets of integers.

* Add cbvrexfw to iset.mm

Stated as in set.mm.  The proof is similar to the iset.mm proof of
cbvralfw .  Also, add some $j usage to related theorems, even though it
doesn't seem like cbvrexfw can have a similar $j usage yet.

* Add nnwofdc and nnwosdc to iset.mm

This is nnwof and nnwos with decidability conditions added.  The proofs
are similar to the set.mm proofs.

* Add infpn to iset.mm

The lemmas infpnlem1 and infpnlem2 are stated as in set.mm.  The proofs
need a small amount of intuitionizing but are basically the set.mm proofs.

The theorem infpn itself is copied without change from set.mm.

* Add prmunb to iset.mm

Stated as in set.mm.  The proof is very similar to the set.mm proof but
needs intuitionizing in one place.

* Add infpn2 and prminf to iset.mm

prminf is stated as in set.mm.  The proof is similar to the set.mm proof but uses
unbendc in place of unben .

infpn2 is stated as in set.mm.  The proof is in terms of prminf because
applying unbendc directly (the way that set.mm applies unben ) would
mean a more involved decidability condition and it is easier to just
take advantage of the decidability already shown as part of prminf .

* "inhabited decidable" in nnwofdc comment

Update the comment to reflect what the theorem says.