Take a closer look at what would be needed to intuitionize 4sq (#4359)
* Add Z[i] to iset.mm

This is the syntax and the df-gz definition.  Copied without change from
set.mm.

* Add Guassian integer theorems to iset.mm

Copied without change from set.mm

This is elgz , gzcn , zgz , igz , gznegcl , gzcjcl , gzaddcl , gzmulcl ,
gzreim , gzsubcl , and gzabssqcl

* Add qsqeqor to iset.mm

This is sqeqor from set.mm but for rational numbers rather than complex
numbers.  The proof is a new one which breaks the proof into four cases
based on whether A and B are nonnegative or nonpositive.

* Add cbvral2vw and cbvrex2vw to iset.mm

Copied from set.mm with the only change being the axioms listed in the
$j usage directive.

* Add 4sqlem5 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in some places but
is basically the set.mm proof.

* Add 4sqlem6 to iset.mm

Stated as in set.mm

The proof needs some intuitionizing but is basically the set.mm proof

* Add 4sqlem7 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* copy 4sqlem8 from set.mm to iset.mm

* copy 4sqlem9 from set.mm to iset.mm

* Add 4sqlem10 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy 4sqlem1 from set.mm to iset.mm

* Add 4sqlem2 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy 4sqlem3 from set.mm to iset.mm

* copy 4sqlem4a from set.mm to iset.mm

* copy 4sqlem4 from set.mm to iset.mm

* Add mul4sqlem to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* copy mul4sq from set.mm to iset.mm

* copy elfzelzd from set.mm to iset.mm

* copy dvds2addd from set.mm to iset.mm

* Update 4sq comments and web pages

Add notes about the 4sq lemmas which we don't prove.  Adjust comments to
not refer to theorems which are not present in iset.mm.