Add Legendre symbol to iset.mm (#4387)
* Add /L to iset.mm

This is the syntax and df-lgs , copied without change from set.mm.

* Copy zabsle1 from iset.mm to set.mm

* Add lgsval to iset.mm

Also lgslem1 , lgslem2 , lgslem3 , and lgslem4 .

All are stated as in set.mm.  The proofs are either unchanged or have
some intuitionizing, but are basically the set.mm proofs.  lgsval needs
the most intuitionizing (to replace ovex with closure theorems).

* Add lgsfvalg to iset.mm

This is lgsfval from set.mm with two additional conditions.  The proof
is broadly similar to the set.mm one but more involved because it
replaces ovex with closure theorems.

* Add lgsfcl2 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various places
but is basically the set.mm proof.

* Add lgscllem to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy lgsfcl and lgsfle1 from set.mm to iset.mm

* Add lgsval2lem to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsval4lem to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy lgscl2 from set.mm to iset.mm

* copy lgs0 and lgscl from set.mm to iset.mm

* copy lgsle1 and lgsval2 from set.mm to iset.mm

* copy lgs2 and lgsval3 from set.mm to iset.mm

* Add lgsvalmod to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Copy lgsval4 from set.mm to iset.mm

* copy lgsfcl3 from set.mm to iset.mm

* Add lgsval4a to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy lgscl1 from set.mm to iset.mm

* Add ifnotdc to iset.mm

This is ifnot from set.mm with a decidability condition added.  The
proof is similar to the set.mm proof.

* Add lgsneg to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsneg1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsmod to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsdilem to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsdir2lem1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy lgsdir2lem2 from set.mm to iset.mm

* copy lgsdir2lem3 from set.mm to iset.mm

* Add lgsdir2lem4 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsdir2lem5 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsdir2 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsdirprm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsdir to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various places
but is basically the set.mm proof.

* Add lgsdilem2 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsdi to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsne0 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing, significant in
places, but is basically the set.mm proof.

* Add lgsabs1 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* copy lgssq from set.mm to iset.mm

* copy lgssq2 from set.mm to iset.mm

* Add lgsprme0 to iset.mm

Stated as in set.mm.  The needs some intuitionizing but is basically the
set.mm proof.

* Add 1lgs to iset.mm

Stated as in set.mm.  The proof requires a very small amount of
intuitionizing but is basically the set.mm proof.

* copy lgs1 from set.mm to iset.mm

* copy lgsmodeq from set.mm to iset.mm

* copy lgsmulsqcoprm from set.mm to iset.mm

* Add lgsdirnn0 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add lgsdinn0 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Remove references to lgsqr in iset.mm comments

We haven't proved it yet.

* Elaborate on class constraints in mmil.html

That is, cases where we need to give a type for a class.  Add some text
about the general situation and be more explicit about how lgsfval is
intuitionized.