Add magmas to iset.mm (#4399)
* Add Mgm and +f to iset.mm

This is the syntax Mgm and +f and the definitions df-mgm and df-plusf .
Copied without change from set.mm.

* Add ismgm and .o. to iset.mm

The .o. variable is copied without change from set.mm.

The ismgm theorem is stated as in set.mm.  The proof needs some
intuitionizing to replace fvexd but this is possible using some
extensible structure theorems.

* Add ismgmn0 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is similar
to the set.mm proof.

* copy ovrspc2v , oveqrspc2v , oveqdr

copy them from set.mm to iset.mm

* copy mgmcl from set.mm to iset.mm

* copy isnmgm from set.mm to iset.mm

* copy mgmsscl from set.mm to iset.mm

Revise the comment to not refer to a theorem we don't have yet.

* Fixes for funexw , mptexw , and mpoexw

Add $j usage to match the axiom usage mentioned in the comments

Copy from set.mm to iset.mm (no changes except as noted here).

Change ax-rep to ax-coll in iset.mm.

* Add plusffvalg to iset.mm

This is plusffval from set.mm with a set existence condition added.  The
proof is similar to the set.mm proof but needs to have various parts
intuitionized.

* Add plusfvalg to iset.mm

This is plusfval with an additional set existence condition.  The proof
needs intuitionizing but is similar to the set.mm proof.

* Add plusfeqg to iset.mm

This is plusfeq from set.mm with an additional set existence condition.
The proof is similar to the set.mm proof.

* Add plusffng to iset.mm

This is plusffn from set.mm with an additional set existence condition.
The proof is similar to the set.mm proof.

* Add mgmplusf to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* add issstrmgm to mmil.html

* Add intopsn to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add mgmb1mgm1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy mgm0 from set.mm to iset.mm

* Add prex and mgm0b to mmil.html

* Add eqtrid to iset.mm

The whole block, including the changes to the syl5eq comment, is copied
without change from set.mm.

* Add mgm1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but generally
follows the set.mm proof.

* Add opifismgmdc to iset.mm

This is opifismgm from set.mm with nonempty changed to inhabited and a
decidability condition added.