Add semigroups to iset.mm (#4414)
* Add gsum theorems to mmil.html

Includes a description of likely changes that would be needed to
intuitionize them (starting with changes to the gsum definition itself).

They are gsumvalx , gsumval , gsumpropd , gsumpropd2lem , gsumpropd2 ,
gsummgmpropd , gsumress , gsumval1 , gsum0 , gsumval2a , gsumval2 ,
gsumsplit1r , gsumprval , and gsumpr12val

* Add Smgrp to iset.mm

This is the syntax and df-sgrp .  Copied without change from set.mm.

* Add issgrp to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy issgrpv from set.mm to iset.mm

* copy issgrpn0 from set.mm to iset.mm

* Add isnsgrp to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy sgrpmgm from set.mm to iset.mm

* copy sgrpass from set.mm to iset.mm

* copy sgrp0 from set.mm to iset.mm

* add sgrp0b to mmil.html

* Add sgrp1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.