Intuitionize section "Definition and basic properties of monoids" (#4453)
* Add Mnd to iset.mm

This is the syntax and df-mnd .  Copied without change from set.mm.

* Add ismnddef to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add ismnd to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add eliota to iset.mm

This is similar to elfv but for any iota expression, not just a
function value.

* Add eliotaeu to iset.mm

This is analogous to relelfvdm but for any iota expression not just
function value.

* Add iotam to iset.mm

This is iotan0 from set.mm with non-empty changed to inhabited.  The
proof is similar to the set.mm proof at least in the sense of being
built on iota2 .

* Add sgrpidmndm to iset.mm

This is sgrpidmnd from set.mm with non-empty changed to inhabited.  The
proof needs some intuitionizing but is based on the set.mm proof.

Remove references in comment to theorems not present in iset.mm

* Copy mndsgrp and mndmgm from set.mm to iset.mm

* copy mndcl and mndass from set.mm to iset.mm

* copy mndid and mndideu from set.mm to iset.mm

* copy mnd32g and mnd12g from set.mm to iset.mm

* copy mnd4g from set.mm to iset.mm

* Add mndidcl and mndbn0 to iset.mm

Copied from set.mm.  The only change is to the comment of mndbn0 ,
to mention non-empty versus inhabited.

* Add fihashelne0d to iset.mm

This is hashelne0d from set.mm but for finite sets.

* Add hashfinmndnn to iset.mm

Stated as in set.mm.  The proof needs a very small amount of intuitionizing
but is other than that the set.mm proof.

* copy mndplusf from set.mm to iset.mm

* Add mndlrid , mndlid , and mndrid to iset.mm

Copied without change from set.mm.

* copy ismndd from set.mm to iset.mm

* Add mndpfo to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add mndfo to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy mndpropd and mndprop from set.mm to iset.mm

* add issubmnd to mmil.html

* Add ress0g and submnd0 to mmil.html

* copy mndinvmod from set.mm to iset.mm

* Add prdsplusgcl to mmil.html

* Add structure product theorems to mmil.html

This is prdsidlem , prdsmndd , prds0g , pwsmnd , and pws0g .

* Add image structure theorems to mmil.html

This is imasmnd2 , imasmnd , and imasmndf1

* add xpsmnd to mmil.html

* Add mnd1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add mnd1id to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.