Intuitionize section "Monoid homomorphisms and submonoids" (#4455)
* Add MndHom and SubMnd to iset.mm

This is the syntax and the definitions df-mhm and df-submnd .

Copied without change from set.mm.

* Add ismhm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy mhmrcl1 and mhmrcl2 from set.mm to iset.mm

* copy mhmf from set.mm to iset.mm

* Add mhmpropd to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy mhmlin from set.mm to iset.mm

* copy mhm0 from set.mm to iset.mm

* copy idmhm from set.mm to iset.mm

* copy mhmf1o from set.mm to iset.mm

* Add issubm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* add issubm2 and issubmndb to mmil.html

* copy issubmd from set.mm to iset.mm

* Add mndissubm to iset.mm

Copied from set.mm except a change to the comment to remove a reference
to a theorem we do not have.

* Add propositional logic theorems to iset.mm

This is anbi1cd , bianass , bianassc , and an21

Copied without change from set.mm.

* add resmndismnd to mmil.html

* Add submss to iset.mm

Stated as in set.mm.  The proof is from submrcl and issubm .

* Add submid to iset.mm

Stated as in set.mm.  The proof is from issubm , mndidcl , and mndcl .

* Add subm0cl to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof but
uses issubm instead of issubm2 .

* copy submcl from set.mm to iset.mm

* Add submmnd , submbas , subm0 to mmil.html

* add subsubm to mmil.html

* Add 0subm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Copy rspc2vd from set.mm to iset.mm

* copy insubm from set.mm to iset.mm

* copy syl2anc2 from set.mm to iset.mm

* Add 0mhm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* add resmhm , resmhm2 , resmhm2b to mmil.html

* copy mhmco from set.mm to iset.mm

* copy mhmima from set.mm to iset.mm

* copy mhmeql from set.mm to iset.mm

* add df-acs and submacs to mmil.html

* add mndind and prdspjmhm to mmil.html

* Add more theorems to mmil.html

This is pwspjmhm , pwsdiagmhm , pwsco1mhm , and pwsco2mhm