Intuitionize section "Group multiple operation" (#4498)
* Add .g to iset.mm

This is the syntax and df-mulg .  Copied without change from set.mm.

* Add mulgfvalg to iset.mm

This is mulgfval from set.mm with a set existence condition added.
The proof is adapted from the mulgfvalALT proof in set.mm and needs some
intuitionizing related to the set existence condition.

* Add $j usage to mulgfval in set.mm

This matches a statement already made in the comment about axiom usage.

* Add ifcli , ifex to mmil.html

* Add mulgval to iset.mm

Stated as in set.mm.  The proof needs a lot of intuitionizing but is
able to follow roughly the outlines of the set.mm proof.

* Add mulgfng to iset.mm

This is mulgfn from set.mm with an additional set existence condition.
The proof is largely taken from the iset.mm proof of mulgval .

* add mulgfvi to mmil.html

* copy mulg0 from set.mm to iset.mm

* copy mulgnn from set.mm to iset.mm

* add mulgnngsum to mmil.html

* add mulgnn0gsum to mmil.html

* Add mulg1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add mulgnnp1 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy mulg2 from set.mm to iset.mm

* copy mulgnegnn from set.mm to iset.mm

* copy mulgnn0p1 from set.mm to iset.mm

* Add mulgnnsubcl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy mulgnn0subcl from set.mm to iset.mm

* copy mulgsubcl from set.mm to iset.mm

* copy mulgnncl from set.mm to iset.mm

* copy mulgnn0cl from set.mm to iset.mm

* copy mulgcl from set.mm to iset.mm

* copy mulgneg from set.mm to iset.mm

* copy mulgnegneg from set.mm to iset.mm

* copy mulgm1 from set.mm to iset.mm

* copy mulgcld from set.mm to iset.mm

* copy mulgaddcomlem from set.mm to iset.mm

* copy mulgaddcom from set.mm to iset.mm

* copy mulginvcom from set.mm to iset.mm

* copy mulginvinv from set.mm to iset.mm

* Add mulgnn0z to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy mulgz from set.mm to iset.mm

* Add mulgnndir to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy mulgnn0dir from set.mm to iset.mm

* copy mulgdir from set.mm to iset.mm

Includes lemma mulgdirlem

* copy mulgp1 from set.mm to iset.mm

* copy mulgneg2 from set.mm to iset.mm

* copy mulgnnass from set.mm to iset.mm

* copy mulgnn0ass from set.mm to iset.mm

* copy mulgass from set.mm to iset.mm

* copy mulgassr from set.mm to iset.mm

* Add mulgmodid to iset.mm

Stated as in set.mm.  The proof needs real->rational modifications in
modulus theorems but is otherwise the set.mm proof.

* copy mulgsubdir from set.mm to iset.mm

* copy mhmmulg from set.mm to iset.mm

* add mulgpropd to mmil.html

* copy submmulgcl from set.mm to iset.mm

* add submmulg to mmil.html

* add pwsmulg to mmil.html

* Correct mulgsubdir comment in set.mm and iset.mm