Add mulGrp to iset.mm (#4509)
* Add lgsqr to mmil.html

* Add df-drng and df-field to mmil.html

Also, revise df-cnfld entry to refer to those items and to slightly
expand the mention of order.

* Add mulGrp to iset.mm

This is the syntax and df-mgp .  It is copied from set.mm with the only
change being to remove from the comment a few references to theorems
iset.mm doesn't have and won't have for a while.

* Add fnmgp to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is based
on the set.mm proof.

* Add mgpvalg to iset.mm

This is mgpval from set.mm with an additional set existence condition.
The proof needs some intuitionizing but is similar to the set.mm proof.

* Add mgpplusgg to iset.mm

This is mgpplusg from set.mm with an additional set existence condition.
The proof needs some intuitionizing but is basically the set.mm proof.

* Add mgpbasg to iset.mm

This is mgpbas from set.mm.  The proof is based on the set.mm proof but
needs intuitionizing.

* add setsplusg to mmil.html

* Add slot theorems to iset.mm

This is scandxnbasendx , scandxnplusgndx , and scandxnmulrndx .
Copied from set.mm.  The only change is to remove comments which refer
to theorems not present in iset.mm.

Update iset-discouraged file.

* Add mgpscag to iset.mm

This is mgpsca from set.mm with a set existence condition added.
The proof is similar to the set.mm proof with some intuitionizing.

* Add tset slot theorems to iset.mm

These are tsetndxnn , basendxlttsetndx , tsetndxnbasendx , tsetndxnplusgndx ,
tsetndxnmulrndx , tsetndxnstarvndx , and slotstnscsi .  They are copied
from set.mm with the only change being to remove comments referring to
theorems not present in iset.mm.

Update iset-discouraged file.

* Add mgptsetg to iset.mm

This is mgptset from set.mm with an additional set existence condition.
The proof needs intuitionizing but is generally similar to the set.mm
proof.

* Add mgptopng to iset.mm

This is mgptopn from set.mm with an additional set existence condition.
The proof needs intuitionizing but is similar to the set.mm proof.

* Add distance slot theorems to iset.mm

This is dsndxnn , basendxltdsndx , dsndxnbasendx , dsndxnplusgndx ,
dsndxnmulrndx , slotsdnscsi , dsndxntsetndx , and slotsdifdsndx .

Copied from set.mm with the only change being to remove references in
the comments to theorems (most of them not in iset.mm).

Update the iset-discouraged file.

* Add mgpdsg to iset.mm

This is mgpds from set.mm with an addition set existence condition.  The
proof is similar to the set.mm proof.

* add mgpress to mmil.html

* Add 1r (ring unit) to iset.mm

This is the syntax and df-ur .  Copied without change from set.mm.

* Add ringidvalg to iset.mm

This is ringidval from set.mm with an additional set existence
condition.  The proof is taken from a portion of the set.mm proof.

* Add dfur2g to iset.mm

This is dfur2 from set.mm with an additional set existence condition.
The proof needs some intuitionizing but is based on the set.mm proof.

* Revise df-ur comment in iset.mm and set.mm

This wording is based on github discussion and is intended to describe
accurately and clearly how this definition works.