Add theorems on domains of addition and multiplication in sigma-algebras (#4516)
* Remove unnecessary $d condition from theorems on sigma-measurable functions

* Add new pre-image theorems

- Reduced  constraints by replacing them with 'effectively not free' hypotheses.
- Improves theorem usability and simplifies certain dependencies.

* Add new theorems involving the |`t operator.

This commit includes:
- Theorems specific to sigma-algebras about the |`t operator.
- Theorems specific to topologies about the |`t operator.
- General theorems about the |`t operator.

Additionally, variations of existing theorems have been added:
- Deduction versions (e.g., ~ inopnd).
- 'Effectively not free' versions (e.g., ~ ss2rabdf).

* Add theorems on domains of addition and multiplication in sigma-algebras

This commit introduces four theorems related to the domains of function addition and multiplication within the context of sigma-algebras:

Theorems adddmmbl and adddmmbl2:
If two functions have domains in a sigma-algebra, the domain of their addition also belongs to the sigma-algebra. These theorems correspond to the first statement of Proposition 121H from [Fremlin1], p. 39. The assumption of sigma-measurability for the functions, as in the original text, is not required here.

Theorems muldmmbl and muldmmbl2:
Similarly, if two functions have domains in a sigma-algebra, the domain of their multiplication also belongs to the sigma-algebra. These theorems correspond to the second statement of Proposition 121H from [Fremlin1], p. 39, again without requiring sigma-measurability.

This commit also adds four missing full stops in previously committed comments.