Add CMnd and Abel to iset.mm (#4514)
* Add CMnd and Abel to iset.mm

This is the syntax, df-cmn , and df-abl .  Copied without change from
set.mm.

* Copy Abelian group theorems to iset.mm

This is isabl , ablgrp , ablgrpd , and ablcmn , copied without change
from set.mm.

* copy iscmn from set.mm to iset.mm

* copy isabl2 from set.mm to iset.mm

* copy cmnpropd from set.mm to iset.mm

* copy ablpropd from set.mm to iset.mm

* copy ablprop from set.mm to iset.mm

* copy iscmnd from set.mm to iset.mm

* copy isabld from set.mm to iset.mm

* copy isabli from set.mm to iset.mm

* copy cmnmnd from set.mm to iset.mm

* copy cmncom from set.mm to iset.mm

* copy ablcom from set.mm to iset.mm

* copy cmn32 from set.mm to iset.mm

* copy cmn4 from set.mm to iset.mm

* copy cmn12 from set.mm to iset.mm

* copy abl32 from set.mm to iset.mm

* copy cmnmndd from set.mm to iset.mm

* copy rinvmod from set.mm to iset.mm

The only change is in the comment, to refer to caovimo instead of caovmo

* copy ablinvadd from set.mm to iset.mm

* copy ablsub2inv from set.mm to iset.mm

* copy ablsubadd from set.mm to iset.mm

* copy ablsub4 from set.mm to iset.mm

* copy abladdsub4 from set.mm to iset.mm

* copy abladdsub from set.mm to iset.mm

* copy ablpncan2 from set.mm to iset.mm

* copy ablpncan3 from set.mm to iset.mm

* copy ablsubsub from set.mm to iset.mm

* copy ablsubsub4 from set.mm to iset.mm

* copy ablpnpcan from set.mm to iset.mm

* copy ablnncan from set.mm to iset.mm

* copy ablsub32 from set.mm to iset.mm

* copy ablnnncan from set.mm to iset.mm

* copy ablnnncan1 from set.mm to iset.mm

* copy ablsubsub23 from set.mm to iset.mm

* Add mulgnn0di to mmil.html

* Add mulgdi to mmil.html

* Adjust commutative monoid comments

Refer to CMnd as a commutative monoid not an Abelian group.  This is in
set.mm and iset.mm for cmn32 , cmn4 , and cmn12 .

* Expand ablpncan2 comment

This is in set.mm and iset.mm.  Although we don't always spell out every
detail in comments, there are enough cancellation laws that it seems
worthwhile to say this is for Abelian groups.

* Say Abelian group is comment for ablpncan3

This is in set.mm and iset.mm

Since those are synonymous (both are, for example, mentioned in the
opening sentence of https://en.wikipedia.org/wiki/Abelian_group ),
I'm a bit reluctant to police which one is used.

But I guess we mostly use Abelian group so :shrug: I guess I can
change this.